Modularize legacy daemon line tests
